AuthRoutes changes
